requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dominate==2.7.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pydantic==1.10.7
python-dotenv==1.0.0
replicate==0.8.1
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
